+++ konst/src/array.rs
@@ -1,12 +1,36 @@
 //! Const equivalents of array functions.
 
+macro_rules! leak_warning {
+    () => {
+        concat!(
+            "# Warning\n",
+            "\n",
+            "This macro leaks the initialized part of the array\n",
+            "if the closure passed to this macro panics or returns early.\n",
+            "\n",
+            "note: this warning is not relevant if the elements don't need dropping",
+            "(e.g: by implementing `Copy`).\n"
+        )
+    };
+}
+
 /// Const equivalent of
 /// [`array::map`](https://doc.rust-lang.org/std/primitive.array.html#method.map).
 ///
-/// **Limitation:** requires `$array` and the elements
-/// returned by the passed-in function to be `Copy`.
+#[doc = leak_warning!()]
+///
+/// # Limitations
+///
+/// This macro supports mapping from non-Copy arrays if any of these
+/// conditions are met about the parameter of the passed-in closure:
+/// 1. it's a pattern that only copies Copy fields of each array element
+/// 2. it's a `ref` pattern
+///
+/// [examples of both of the above conditions below](#map-noncopy-example)
 ///
 /// # Example
+///
+/// ### Basic
 ///
 /// ```rust
 /// use konst::array;
@@ -28,13 +52,41 @@
 ///
 /// ```
 ///
+/// <span id="map-noncopy-example"> </span>
+/// ### Map from non-Copy array
+///
+/// Demonstrates both ways to map from a non-Copy array.
+///
+/// ```rust
+/// use konst::array;
+///
+/// struct NonCopy(u32, u32);
+///
+/// const PRIME_SUMS: [u32; 3] = {
+///     let input = [NonCopy(2, 3), NonCopy(5, 7), NonCopy(11, 13)];
+///     
+///     // demonstrates the first way to map from non-Copy elements
+///     array::map!(input, |NonCopy(l, r)| l + r)
+/// };
+/// assert_eq!(PRIME_SUMS, [5, 12, 24]);
+///
+/// const FIBB_SUMS: [u32; 3] = {
+///     let input = [NonCopy(2, 3), NonCopy(5, 8), NonCopy(13, 21)];
+///     
+///     // demonstrates the second way to map from non-Copy elements
+///     array::map!(input, |ref nc| nc.0 + nc.1)
+/// };
+/// assert_eq!(FIBB_SUMS, [5, 13, 34]);
+///
+/// ```
+///
 pub use konst_kernel::array_map as map;
 
 /// Const equivalent of [`array::from_fn`](core::array::from_fn).
 ///
-/// # Limitations
+#[doc = leak_warning!()]
 ///
-/// This macro requires the element type to be `Copy`.
+/// # Limitations
 ///
 /// When the array type is annotated, the array type must be one of:
 /// - Square brackets (e.g: `from_fn!([usize; 10] => |i| i)`)