+++ konst_kernel/src/macros/array_macros.rs
@@ -7,7 +7,7 @@ macro_rules! array_map {
             ref array => {
                 let array = $crate::__::assert_array(array);
 
-                $crate::utils::__alt_parse_closure_1!{
+                $crate::utils::__parse_closure_1!{
                     ($crate::__array_map) (array, |i| array[i],) (array_map),
                     $($closure)*
                 }
@@ -22,18 +22,18 @@ macro_rules! __array_map {
     (
         $array:ident,
         |$i:ident| $get_input:expr,
-        |$param:tt $(: $type:ty)? $(,)?| $(-> $ret:ty)? $mapper:block $(,)?
+        ($($pattern:tt)*) $(-> $ret:ty)? $mapper:block $(,)?
     ) => ({
         let len = $array.len();
-        let mut out = $crate::__::uninit_copy_array_of_len(&$array);
+        let mut out = $crate::__::uninit_array_of_len(&$array);
 
         let mut $i = 0usize;
         while $i < len {
-            let $param $(: $type)? = $get_input;
+            let $($pattern)* = $get_input;
             out[$i] = $crate::__::MaybeUninit $(::<$ret>)? ::new($mapper);
             $i += 1;
         }
-        // protecting against malicious `$mapper`s
+        // protecting against malicious `$mapper`s that break out of the `while` loop
         $crate::__::assert!($i == len);
 
         unsafe{
@@ -49,7 +49,7 @@ macro_rules! array_from_fn {
     ($type:tt => $($closure:tt)*) => ({
         let input = $crate::__::unit_array();
         let arr: $crate::__unparenthesize!($type) =
-            $crate::utils::__alt_parse_closure_1!{
+            $crate::utils::__parse_closure_1!{
                 ($crate::__array_map) (input, |i| i,) (array_from_fn),
                 $($closure)*
             };
@@ -57,7 +57,7 @@ macro_rules! array_from_fn {
     });
     ($($closure:tt)*) => ({
         let input = $crate::__::unit_array();
-        $crate::utils::__alt_parse_closure_1!{
+        $crate::utils::__parse_closure_1!{
             ($crate::__array_map) (input, |i| i,) (array_from_fn),
             $($closure)*
         }
@@ -72,10 +72,7 @@ pub const fn assert_array<T, const N: usize>(array: &[T; N]) -> &[T; N] {
 }
 
 #[inline(always)]
-pub const fn uninit_copy_array_of_len<T, U, const N: usize>(_input: &[T; N]) -> [MaybeUninit<U>; N]
-where
-    U: Copy,
-{
+pub const fn uninit_array_of_len<T, U, const N: usize>(_input: &[T; N]) -> [MaybeUninit<U>; N] {
     crate::maybe_uninit::uninit_array()
 }